Add days-to-expiration to option quotes
Also allows a nicer format for option quotes where we show the
symbol broken out on the right with days to expiration.

TODO: make the days-to-expiration calculation more adaptive so it
shrinks down to "hours to expiration" then "minutes" then "seconds" as
actual deadlines approach.

Also note this changes the default time collector to always pin
US/Eastern because that's what we use everywhere else (and the "days to
expiration" math requires matching timezones).